lab9/requirements.txt
Pillow==9.5.0
numpy==1.21.6
matplotlib==3.5.3
